Segregate junit4 dependencies to allow using with cucumber-junit-platform-engine and junit-platform-suite
