+++ serenity-cucumber/src/main/java/net/serenitybdd/cucumber/model/StoredFeatureFile.java
@@ -1,6 +1,6 @@
 package net.serenitybdd.cucumber.model;
 
-import net.serenitybdd.cucumber.CucumberWithSerenity;
+import net.serenitybdd.cucumber.CucumberWithSerenityRuntime;
 
 import java.io.File;
 import java.io.IOException;
@@ -47,7 +47,7 @@ public class StoredFeatureFile {
     
 
     public File fromTheConfiguredPaths() throws IOException {
-        for(URI uri : CucumberWithSerenity.currentRuntimeOptions().getFeaturePaths()) {
+        for(URI uri : CucumberWithSerenityRuntime.currentRuntimeOptions().getFeaturePaths()) {
             if (Files.exists(candidatePath(uri, featureFileName))) {
                 return candidatePath(uri, featureFileName).toFile();
             }
